Updated images for 1989 observing campaign of 28 Sgr’s occultation by Saturn’s rings (#84)
* Add RMS card images

Uploaded card images for the investigations that RMS is responsible for.

* Add RMS hero images

Uploaded card images for the investigations that RMS is responsible for.

* Update RMS hero images

This update replaces the placeholder image for the 1989 observing campaign of 28 Sgr's occultation by Saturn's rings.

* Update RMS card images

This update replaces the placeholder image for the 1989 observing campaign of 28 Sgr's occultation by Saturn's rings.

* Delete data/pds4/context_support/images/card/investigation/PLACEHOLDER_observing_campaign.saturn_occultation_of_28_sagittarius_1989.card.png

* Delete data/pds4/context_support/images/hero/investigation/PLACEHOLDER_observing_campaign.saturn_occultation_of_28_sagittarius_1989.hero.png

---------

Co-authored-by: Megan Seritan <meganseritan@gmail.com>