Update repo to only track latest (#94)
* Update repo to only track latest

To avoid some of the inconsistency we are seeing between the archive and the EdWG versions of context products, we are going to move to a new paradigm of only maintaining the latest version of context products in this repo.

All past versions will be maintained in the archive.

More process improvements will follow to ensure we properly version these products moving forward.

* Update missing agencies

* removed *_to_agency; removed telescope_to_telescope

* Rename context_support to context-support

---------

Co-authored-by: rchenatjpl <rchen@jpl.nasa.gov>